+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.2.0
 
 - Add `sort_attrs` method in `Markup` class.
+- Add `merge_attrs` method in `Markup` class.
 
 ## 3.1.0